requirements.txt
streamlit
groq
pinecone-client
fastembed
